+++ test/index.test.js
@@ -1,6 +1,5 @@
 "use strict";
 
-const fs = require("fs");
 const { default: jsTokens } = require("../");
 
 function token(name, fn) {
@@ -322,7 +321,8 @@ describe("tokens", () => {
     match("/a/gmiyus");
     match("/a/myg");
     match("/a/e");
-    match("/a/invalidFlags", "/a/invalid");
+    match("/a/E");
+    match("/a/invalidFlags");
     match("/a/f00", "/a/f");
 
     match("/\n/", false);
@@ -799,6 +799,7 @@ describe("tokens", () => {
 
     match("(1/2)/g", ["(", "1", "/", "2", ")", "/", "g"]);
     match("(x)/a/g", ["(", "x", ")", "/", "a", "/", "g"]);
+    match("+{}/a/g", ["+", "{", "}", "/", "a", "/", "g"]);
     match("5./a/g", ["5.", "/", "a", "/", "g"]);
     match("x++/a/g", ["x", "++", "/", "a", "/", "g"]);
     match("x--/a/g", ["x", "--", "/", "a", "/", "g"]);
@@ -815,183 +816,4 @@ describe("tokens", () => {
     match("☃");
     match("💩");
   });
-});
-
-test("switch", () => {
-  function token(match) {
-    const value = match[0];
-
-    switch (value) {
-      case match.groups.StringLiteral:
-        return {
-          type: "StringLiteral",
-          value,
-          closed: match.groups.StringLiteralEnd !== undefined,
-        };
-
-      case match.groups.Template:
-        return {
-          type: "Template",
-          value,
-          closed: match.groups.TemplateEnd !== undefined,
-        };
-
-      case match.groups.MultiLineComment:
-        return {
-          type: "MultiLineComment",
-          value,
-          closed: match.groups.MultiLineCommentEnd !== undefined,
-        };
-
-      case match.groups.SingleLineComment:
-        return { type: "SingleLineComment", value };
-
-      case match.groups.RegularExpressionLiteral:
-        return { type: "RegularExpressionLiteral", value };
-
-      case match.groups.NumericLiteral:
-        return { type: "NumericLiteral", value };
-
-      case match.groups.IdentifierName:
-        return { type: "IdentifierName", value };
-
-      case match.groups.Punctuator:
-        return { type: "Punctuator", value };
-
-      case match.groups.WhiteSpace:
-        return { type: "WhiteSpace", value };
-
-      case match.groups.LineTerminatorSequence:
-        return { type: "LineTerminatorSequence", value };
-
-      case match.groups.Invalid:
-        return { type: "Invalid", value };
-
-      default:
-        throw new Error("Should never be reached");
-    }
-  }
-
-  const code = 'console.log("", ``, /**/, /./, 0x1Fn) //\r\n#\'';
-
-  const tokens = Array.from(code.matchAll(jsTokens)).map(token);
-
-  expect(tokens).toMatchInlineSnapshot(`
-    Array [
-      Object {
-        "type": "IdentifierName",
-        "value": "console",
-      },
-      Object {
-        "type": "Punctuator",
-        "value": ".",
-      },
-      Object {
-        "type": "IdentifierName",
-        "value": "log",
-      },
-      Object {
-        "type": "Punctuator",
-        "value": "(",
-      },
-      Object {
-        "closed": true,
-        "type": "StringLiteral",
-        "value": "\\"\\"",
-      },
-      Object {
-        "type": "Punctuator",
-        "value": ",",
-      },
-      Object {
-        "type": "WhiteSpace",
-        "value": " ",
-      },
-      Object {
-        "closed": true,
-        "type": "Template",
-        "value": "\`\`",
-      },
-      Object {
-        "type": "Punctuator",
-        "value": ",",
-      },
-      Object {
-        "type": "WhiteSpace",
-        "value": " ",
-      },
-      Object {
-        "closed": true,
-        "type": "MultiLineComment",
-        "value": "/**/",
-      },
-      Object {
-        "type": "Punctuator",
-        "value": ",",
-      },
-      Object {
-        "type": "WhiteSpace",
-        "value": " ",
-      },
-      Object {
-        "type": "RegularExpressionLiteral",
-        "value": "/./",
-      },
-      Object {
-        "type": "Punctuator",
-        "value": ",",
-      },
-      Object {
-        "type": "WhiteSpace",
-        "value": " ",
-      },
-      Object {
-        "type": "NumericLiteral",
-        "value": "0x1Fn",
-      },
-      Object {
-        "type": "Punctuator",
-        "value": ")",
-      },
-      Object {
-        "type": "WhiteSpace",
-        "value": " ",
-      },
-      Object {
-        "type": "SingleLineComment",
-        "value": "//",
-      },
-      Object {
-        "type": "LineTerminatorSequence",
-        "value": "
-    ",
-      },
-      Object {
-        "type": "Invalid",
-        "value": "#",
-      },
-      Object {
-        "closed": false,
-        "type": "StringLiteral",
-        "value": "'",
-      },
-    ]
-  `);
-});
-
-describe("tokenization", () => {
-  function testFile(file) {
-    const contents = fs.readFileSync(`test/fixtures/${file}.js`).toString();
-    const expected = require(`./fixtures/${file}.json`);
-    const actual = contents.match(jsTokens);
-    test(`${file}.js`, () => {
-      expect(actual).toEqual(expected);
-      expect(actual.join("")).toBe(contents);
-    });
-  }
-
-  testFile("base64");
-  testFile("errors");
-  testFile("regex");
-  testFile("division");
 });
